Introduce DI with Guice
Add modules for core code and infrastructure

Add a nice junit rule that allows us to wire up dependencies in unit tests
as well.

Unit test uses a mock module

*An interesting effect of the stateful method of unit testing is that
the test still passes even without the mock because it returns the exact
same response as the actual web service.